Improving txFee calculations, fixes for CuraDAI
