+++ server/src/tasks/index.js
@@ -2,7 +2,6 @@ const deployment = require('./deployment')
 const events = require('./events')
 const actions = require('./actions')
 const alert = require('./alert')
-const wallets = require('./wallets')
 const tvl = require('./tvl')
 
 module.exports = {
@@ -10,6 +9,5 @@ module.exports = {
   ...deployment,
   ...actions,
   ...alert,
-  ...wallets,
   ...tvl
 }
